点赞模块 - 优化从redis中读

+++ src/main/java/com/easen/aicode/model/vo/AppThumbDetailVO.java
@@ -59,6 +59,6 @@ public class AppThumbDetailVO implements Serializable {
     /**
      * 点赞数
      */
-    private Long thumbCount;
+    private Integer thumbCount;
 
 }
